rename a file for better coherence
