+++ x-pack/test/security_solution_cypress/cypress/urls/rule_edit.ts
